// src/types/generated-admin-types.ts
export type Maybe<T> = T | null;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]?: Maybe<T[SubKey]> };
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]: Maybe<T[SubKey]> };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  Date: any;
  /** A date-time string at UTC, such as 2007-12-03T10:15:30Z, compliant with the `date-time` format outlined in section 5.6 of the RFC 3339 profile of the ISO 8601 standard for representation of dates and times using the Gregorian calendar. */
  DateTime: any;
  /** The `JSON` scalar type represents JSON values as specified by [ECMA-404](http://www.ecma-international.org/publications/files/ECMA-ST/ECMA-404.pdf). */
  JSON: any;
  /** The `Upload` scalar type represents a file upload. */
  Upload: any;
};
// ...
export type CreateNfeCompanyInput = {
  ownerName: Scalars['String'];
  corporateName: Scalars['String'];
  email: Scalars['String'];
  phone: Scalars['String'];
  tradeName?: Maybe<Scalars['String']>;
  accounting?: Maybe<Scalars['String']>;
  subdomain?: Maybe<Scalars['String']>;
  logoUrl?: Maybe<Scalars['String']>;
  cpf: Scalars['String'];
  cnpj: Scalars['String'];
  ie: Scalars['String'];
  unit: NfeCompanyUnit;
  address: NfeCompanyAddressInput;
  taxSettings?: Maybe<NfeCompanyTaxSettingsInput>;
  certificate?: Maybe<NfeCompanyCertificateInput>;
  emitterCredentials?: Maybe<NfeCompanyEmitterCredentialsInput>;
};
// ...
export type NfeCompanyAddress = {
  __typename?: 'NfeCompanyAddress';
  postcode: Scalars['String'];
  street: Scalars['String'];
  number: Scalars['String'];
  complement?: Maybe<Scalars['String']>;
  neighborhood: Scalars['String'];
  city: Scalars['String'];
  state: Scalars['String'];
};

export type NfeCompanyAddressInput = {
  postcode: Scalars['String'];
  street: Scalars['String'];
  number: Scalars['String'];
  complement: Scalars['String'];
  neighborhood: Scalars['String'];
  city: Scalars['String'];
  state: Scalars['String'];
};

export type NfeCompanyCertificate = {
  __typename?: 'NfeCompanyCertificate';
  base64?: Maybe<Scalars['String']>;
  password?: Maybe<Scalars['String']>;
};

export type NfeCompanyCertificateInput = {
  base64?: Maybe<Scalars['String']>;
  password?: Maybe<Scalars['String']>;
};

export type NfeCompanyEmitterCredentials = {
  __typename?: 'NfeCompanyEmitterCredentials';
  consumer_key?: Maybe<Scalars['String']>;
  consumer_secret?: Maybe<Scalars['String']>;
  access_token?: Maybe<Scalars['String']>;
  access_token_secret?: Maybe<Scalars['String']>;
};

export type NfeCompanyEmitterCredentialsInput = {
  consumer_key?: Maybe<Scalars['String']>;
  consumer_secret?: Maybe<Scalars['String']>;
  access_token?: Maybe<Scalars['String']>;
  access_token_secret?: Maybe<Scalars['String']>;
};
// ...
export type NfeCompanyTaxSettingsInput = {
  ibpt?: Maybe<Scalars['String']>;
  informacoes_fisco?: Maybe<Scalars['String']>;
  orientacao_danfe?: Maybe<Scalars['String']>;
  nfe_serie?: Maybe<Scalars['Int']>;
  nfe_numero?: Maybe<Scalars['Int']>;
  nfe_numero_dev?: Maybe<Scalars['Int']>;
  nfce_serie?: Maybe<Scalars['Int']>;
  nfce_numero?: Maybe<Scalars['Int']>;
  nfce_numero_dev?: Maybe<Scalars['String']>;
  nfce_id_csc?: Maybe<Scalars['String']>;
  nfce_codigo_csc?: Maybe<Scalars['String']>;
  nfce_id_csc_dev?: Maybe<Scalars['String']>;
  nfce_codigo_csc_dev?: Maybe<Scalars['String']>;
};

export enum NfeCompanyUnit {
  Matriz = 'matriz',
  Filial = 'filial'
}
// ...
export type UpdateNfeCompanyInput = {
  id: Scalars['ID'];
  ownerName: Scalars['String'];
  corporateName: Scalars['String'];
  email: Scalars['String'];
  phone: Scalars['String'];
  tradeName?: Maybe<Scalars['String']>;
  accounting?: Maybe<Scalars['String']>;
  subdomain?: Maybe<Scalars['String']>;
  logoUrl?: Maybe<Scalars['String']>;
  cpf: Scalars['String'];
  cnpj: Scalars['String'];
  ie: Scalars['String'];
  unit: NfeCompanyUnit;
  address: NfeCompanyAddressInput;
  taxSettings?: Maybe<NfeCompanyTaxSettingsInput>;
  certificate?: Maybe<NfeCompanyCertificateInput>;
  emitterCredentials?: Maybe<NfeCompanyEmitterCredentialsInput>;
};